+++ R/define_cases_and_controls.R
@@ -29,16 +29,16 @@ define_cases_and_controls = function(dataset)
       `FX3*` =
         case_when(
           `CGG` < 55 ~ "CGG < 55",
-          `CGG` |> between(55, 100) ~ "CGG 55-100",
-          `CGG` |>  between(101, 200) ~ "CGG 101-200",
-          `CGG` > 200 ~ "CGG > 200",
+          `CGG` |> between(55, 99) ~ "CGG 55-99",
+          `CGG` |>  between(100, 199) ~ "CGG 101-199",
+          `CGG` >= 200 ~ "CGG >= 200",
           is.na(`CGG`) ~ "CGG missing",
           .ptype = factor(
             levels = c(
               "CGG < 55",
-              "CGG 55-100",
-              "CGG 101-200",
-              "CGG > 200",
+              "CGG 55-99",
+              "CGG 100-199",
+              "CGG >= 200",
               "CGG missing")
           )
         )